pyproject.toml
[tool.poetry]
name = "tasktracker-django"
version = "1.0.1"
description = "the demo django project - todo application"
authors = ["Serhii Horodilov <sgorodil@gmail.com>"]
license = "MIT"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.1"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2023.10"
pillow = "^10.2.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
